LayoutTests/platform/blackberry/editing/deleting/5408255-expected.txt

CONSOLE MESSAGE: line 18: TypeError: 'null' is not an object (evaluating 'deleteButton.offsetParent')
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     "

"
|     <div>
|       contenteditable="true"
|       id="test"
|       style="padding: 1em;"
|       "
"
|       <ul>
|         class="needsDeletionUI"
|         <li>
|           "1"
|         <li>
|           id="li"
|           "<#selection-caret>2"
|       "
"
|     "

"
|     <script>
|       src="../../resources/dump-as-markup.js"
|     "
"
|     <script>
|       "

sel = window.getSelection();
li = document.getElementById("li");
sel.setPosition(li, 0);

if (window.testRunner) {
    deleteButton = document.getElementById("WebKit-Editing-Delete-Button");
    x = deleteButton.offsetParent.offsetLeft + deleteButton.offsetParent.offsetParent.offsetLeft + deleteButton.offsetLeft + deleteButton.offsetWidth / 2;
    y = deleteButton.offsetParent.offsetTop + deleteButton.offsetParent.offsetParent.offsetTop + deleteButton.offsetTop + deleteButton.offsetHeight / 2;
    eventSender.mouseMoveTo(x, y);
    eventSender.mouseDown();
    eventSender.mouseUp();
    deleteButton = document.getElementById("WebKit-Editing-Delete-Button");
    testContainer = document.getElementById("test");
    Markup.description("There should be no visible content in the markup below. This test is for a bug where the delete button wouldn't work because it had -webkit-user-select:none instead of -webkit-user-select:ignore.");
    if (deleteButton)
        testContainer.innerHTML += "FAIL";
    Markup.dump("test");
}

"
|     "


"
